Refactor Session to remove code duplication
There was a lot of code duplication involved in just juggling and
handing off various fields of `ValidatorConfig`.

Instead of that, this makes a copy of the `ValidatorConfig` at the time
a given client thread is started, and passes is it through wholesale,
storing it in Session.

Finally, `Connection` is made into a proper trait, allowing the generic
parameter on `Session` to be removed, instead storing the session in
`Box<dyn Connection>`.

This massively simplifies the involved code.