chain/state: Avoid panicking in update_consensus_state
Previously this codepath assumed the `ConsensusState` `block_ids`are
`Some(...)` as the logic above used to explicitly check for it.

This was recently changed to allow them to be `None` (a.k.a. `<nil>`),
causing a panic in the event they are `None` (and a `PoisonError`,
since they're holding a mutex on the chain state.

This uses a helper method which handles both cases and displays `<nil>`
in the event they're none.