Detect and don't attempt to recover from PoisonError
PoisonError indicates that a mutex-guarded data structure is in an
inconsistent state because a panic occurred while the mutex was held.

We can't recover from these, so detect when they happen and exit
afterward.

Additionally this names each client thread to ease debugging, and adds
some better debug logging in several places.