Double signing detection and logging improvements
It appears the double signing detection wasn't working correctly in
cases related to round step changes because the round step is presently
hardcoded to `6` in the Amino parser:

https://github.com/interchainio/tendermint-rs/blob/566dfb6/src/amino_types/vote.rs#L170

We have the step information in the form of the Amino `SignedMsgType`:

https://docs.rs/tendermint/0.10.0/tendermint/amino_types/signature/enum.SignedMsgType.html

This commit adds a `parse_request` method which sets the step number to
the following mapping suggested by @bucky:

- Proposal => 0
- PreVote => 1
- PreCommit => 2

(Note: Ideally this would get fixed upstream in `tendermint-rs`)

Additionally, it changes the double signing detection logic to permit a
case which was previously disallowed, which is a `PreVote` for a
particular block ID followed by a `PreCommit` for `<nil>`. This case was
added as a new test to the existing cases, and now passes.