Vendor Secret Connection (fixes interchainio/tendermint-rs#27)
This commit wires up the Secret Connection implementation, whose code is
otherwise by definition identical to the `tendermint-rs` v0.10 crate
today, as it used to live in this repo and was restored by reverting it
in commit 104bb829.

Otherwise this is the minimum necessary changes needed to use the
in-tree copy as opposed to `tendermint-rs`. It otherwise performs no
changes or upgrades other than the ones needed to make the code compile
and pass the latest `clippy`.