Upgrade signatory -> 0.15; yubihsm -> 0.28; use chacha20poly1305 crate
This commit updates `signatory` and `yubihsm` to the latest versions.

This requires either upgrading *ring* to 0.16, or replacing it:

https://github.com/interchainio/tendermint-rs/issues/11

This commit has opted to replace it with the (i.e. my)
`chacha20poly1305` crate:

https://github.com/RustCrypto/AEADs/tree/master/chacha20poly1305

While *ring* is a great crate with excellent cryptographic
implementations, the inability to link two versions of `ring-asm` is a
major impediment to doing upgrades, as it requires updating all
*ring*-dependent dependencies at once, or not at all, which so far has
been a major blocker and has been stalling KMS development for awhile.
As we're only using it for ChaCha20Poly1305, it seems like a rather
heavyweight dependency just to serve that one purpose.