Validate chains are registered on startup
Previously chains were checked lazily at the time their pubkeys were
fetched or signing was requested.

This adds an eager check at the time clients start.

It would probably be better if this check mapped to some sort of
known-good handle (e.g. `Index` into a generational arena), but this
is enough to catch the error eagerly instead of lazily.

I encountered this personally during the `cosmoshub-3` upgrade.