Upgrade `x25519-dalek` to v0.6; remove `rand_os`
The `rand_os` crate is deprecated and uses an out-of-date `rand_core`
dependency, which was previously blocking the `x25519-dalek` upgrade.

This replaces it with `rand` v0.7 and `rand::thread_rng()`, a
`CryptoRng` (ensured by dalek's trait bounds) seeded by the OS RNG.

For producing the YubiHSM master secret used during setup, the
`getrandom` crate is used directly, since this is guaranteed to be a
direct interface to the OS RNG (it also pulls randomness from the
YubiHSM itself).