Replace `lazy_static` with `once_cell`
The `once_cell` crate provides a macro-free alternative to `lazy_static`
and is already used by Abscissa.

There's also an open RFC to incorporate `once_cell` into the Rust
standard library:

https://github.com/rust-lang/rfcs/pull/2788

This PR does a straightforward replacement of `lazy_static` with
`once_cell::sync::Lazy`.

However some of this code (particularly the YubiHSM initialization code)
is dirtier than it needs to be because of the `lazy_static` API, and
would benefit from refatcoring to use `once_cell::sync::OnceCell`
instead, which allows a one-time initialization.

Notably this would be useful for replacing the the `init_connector()`
and `init_client()` methods (as well as the `CLI_COMMAND` static),
which are all workarounds for a lazy initialization-based API.