tx_signer: add "JSONRPC" request params; refactor response
This commit sends a JSON-serialized body of parameters with each request
to a "transaction source" (i.e. oracle):

- Chain ID
- Arbitrary context string (optional; taken from tmkms config)
- Last Tendermint RPC `/status` response
- Commit response from last broadcasted transaction

This should make it possible for Delphi to detect that an error occured
broadcasting the last transaction and skip voting during that period.

Additionally it makes a somewhat breaking change to the response format
expected from a transaction source: it now (optionally) expects a single
transaction, rather than a vec of them. This simplifies the response
error handling, and encourages sources to pack more messages into a
single transaction rather than using multiple transactions.