tx_signer: speculatively increment sequence on error
When encountering an error broadcasting a transaction, speculatively
retry the operation once with an incremented sequence number.

This is to handle the case where a transaction was successfully
broadcast, but an error occurred returning the response.

In such a case, the sequence number may be out-of-sync with the chain,
so if we replay the transaction in such cases we may get it to succeed.