tx_signer: fix retry-on-error behavior
Fixes the logic to retry failed transactions at a higher sequence number
if the previous attempts to broadcast a transaction resulted in an
error.

Prior to this commit, the logic was never triggered as it needed to
store whether or not the prior transaction succeeded before performing a
new one.