secret_connection: refactor in prep for Tendermint v0.34 work
This commit contains a set of initial refactoring to better support
multiple versions of Secret Connection.

It leans more heavily on the (newly extracted) `protocol::Version` enum,
moving much of the protocol framing there in advance so as to lean on
the enum itself to decide how messages are framed.