tx-signer: add a third retry for failed transactions (#213)
When we encounter an error broadcasting a transaction, we don't know if
it actually went on to succeed or not. If it did, we need to increment
the sequence number.

Ideally we could simply query the chain for either the transaction to
see if it was actually broadcast or to simply check the current sequence
number so we could resynchronize in this case, but we can't do that
until after Tendermint v0.34-related protobuf upgrades.

In the meantime, this commit puts a band-aid on a band-aid and adds a
third retry in these scenarios, which are happening several times a
week.

If this doesn't address it, the retry logic could be extended into a
loop, but that gets tricky.