Rename and refactor `SignableMsg::signable_bytes`
The previous method name was `sign_bytes`, which as was noted in #273 is
unclear to whether the resulting bytes have an appended signature.

The new name indicates that the message can be signed, but has not yet
been signed.

Additionally, this commit does a bit of refactoring to eliminate some
duplicated code within the method.

Closes #273