Add `SignableMsg::add_*_signature`; log extension sigs
Moves the methods for adding signatures to a `SignableMsg` onto the
type itself:

- `SignableMsg::add_consensus_signature`
- `SignableMsg::add_extension_signature`

This helps keep as much of the signature-related logic on the
`SignableMsg` type itself, rather than spread out elsewhere.

Also, adds logging every time an extension signature is computed, so we
can confirm it isn't being computed for chains which aren't using it.