Add optional `sign_extensions` field to `[[chain]]` config
This adds a new field to the `ChainConfig` structure which represents an
entry in the `[[chain]]` registry in tmkms.toml which indicates whether
or not vote extensions should be signed.

In the future we should be able to query this using the `signExtension`
field of `SignVote`. See cometbft/cometbft#2439.

However, for now this requires a manual configuration flag.