tmkms-p2p: import from `tendermint-rs`
We will be migrating to `cometbft-rs`, which does not include a P2P
crate, i.e. `tendermint-p2p` is effectively EOL.

This migrates the `tendermint-p2p` crate (back) in-tree as the
`tmkms-p2p` crate for now. This will make it possible to migrate the
crate's dependencies to their `cometbft` equivalents, as well as
upgrading the crypto dependencies when/where needed.

Sources are imported as of this commit:

    commit d8c361ec5597e3f9a005d7211eec9c74eb9b04df
    Author: Romain Ruetschi <romain@informal.systems>
    Date:   Mon May 5 11:36:26 2025 +0200

    Release v0.40.4 (#1503)