tmkms-p2p: reject low order X25519 outputs
This is a bit belt-and-suspenders since the real security is via the
Merlin transcript, however this replaces the old check for the X25519
output being zero with one that checks all low order points.

Additionally, it makes the `reject_low_order_points` function
constant-time, since it's now working with a secret.

Finally, it adds a very basic test that it works.