tmkms-p2p: add `ReadFrame` trait
Adds a trait for moving buffer allocation for frames to `tmkms-p2p`,
rather than having tmkms allocate the buffer.

This is a bit awkward and it would be nice if it could just be an
inherent method on `SecretConnection` instead, but this fits within the
somewhat awkward abstractions that are currently in place to support
Unix sockets in addition to SecretConnection.

(Ideally we'd just take separate I/O paths for SecretConnection vs Unix
sockets, or just drop support for Unix sockets entirely)

Uses `Bytes` as the return type, introducing `bytes` as a dependency.
This is more interesting for forthcoming async support, but we can
use it in both places.