tmkms-p2p: remove explicit low order point check
Instead test how the implementation responds to low-order points.

Previously there were *three* redundant mitigations for the original
MitM attack[1] on the Secret Connection protocol:

1. Eager check that peer's ephemeral DH key isn't a low-order point
2. Post-DH check that the output is non-zero
3. Merlin transcript (which actually provides AKE-ish security)

Looking at the Noise Protocol as a similar but well-designed protocol,
its security rests in the equivalent of 3. 2 is allowed, but a "MAY",
and not actually needed for security (but would independently mitigate
the MitM attack from [1])

[1]: https://eprint.iacr.org/2019/526.pdf