tmkms-p2p: fewer `Frame` panics
`Frame` is a type which is right at the network edge, so any panicking
APIs invite a potential DoS in the event of bugs.

This commit converts potential panics into runtime errors which will
abort the individual connection rather than the whole process.

It also gets all of the I/O-related code out of `Frame`, allowing it to
be purely an encryption-related type.