Add `protocol_version = "v0.38"`
We'll need to migrate to CometBFT V1 at some point, which contains
breaking changes.

Rather than getting rid of the `protocol_version` field under
`[validator]`, it can instead be repurposed to assist in the migration.

First we need to get people to update it to `v0.38`, which is the proto
version we source from (iq-)`cometbft-rs`.

After that we can add support for `v1`, and eventually make it the
default.